requirements.txt
pdf2image>=1.17.0
python-pptx>=0.6.23
Pillow>=10.0.0
google-generativeai>=0.8.0
openai>=1.60.0
anthropic>=0.45.0
pymupdf>=1.25.0
python-dotenv>=1.0.0
rich>=13.9.0
